refactor(lotto): getWinningNumbers를 WinningNumbers 의 정적팩토리 메서드로 변경
