+++ src/main/kotlin/lotto/WinningNumbers.kt
@@ -1,3 +1,19 @@
 package lotto
 
-class WinningNumbers(val winningNumbers: List<Int>)
+class WinningNumbers(val winningNumbers: List<Int>) {
+    companion object {
+        fun getInstance(winningNumbers: String): WinningNumbers {
+            val numbers = winningNumbers.replace(" ", "")
+                .split(InputView.DELIMITER)
+                .map { it.toInt() }
+            validateDuplicatedNumber(numbers)
+            return WinningNumbers(numbers)
+        }
+
+        private fun validateDuplicatedNumber(numbers: List<Int>) {
+            if (numbers.distinct().size != numbers.size) {
+                throw IllegalArgumentException("중복된 숫자는 입력할 수 없습니다.")
+            }
+        }
+    }
+}
